+++ src/GRA.Controllers/ViewModel/Profile/ProfileDetailViewModel.cs
@@ -19,6 +19,8 @@ namespace GRA.Controllers.ViewModel.Profile
         public string SystemName { get; set; }
         public string BranchName { get; set; }
 
+        public bool AskEmailReminder { get; set; }
+
         public string TranslationDescriptionPastTense { get; set; }
         public string ActivityDescriptionPlural { get; set; }
     }