+++ src/GRA.Data/Model/User.cs
@@ -59,6 +59,7 @@ namespace GRA.Data.Model
         public DateTime? LastBroadcast { get; set; }
 
         public bool IsFirstTime { get; set; }
+        public bool PreregistrationReminderRequested { get; set; }
 
         public int? DailyPersonalGoal { get; set; }
     }