+++ src/GRA.Domain.Model/User.cs
@@ -82,6 +82,8 @@ namespace GRA.Domain.Model
 
         [DisplayName("Is this your first time participating?")]
         public bool IsFirstTime { get; set; }
+        [DisplayName("Would you like for us to send an email reminder when the program starts?")]
+        public bool PreregistrationReminderRequested { get; set; }
         [DisplayName("Set a personal goal")]
         public int? DailyPersonalGoal { get; set; }
     }